+++ src/main/kotlin/blackjack/domain/participant/BetAmount.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.domain.participant
 
 @JvmInline
 value class BetAmount(val value: Int) {
