+++ src/main/kotlin/blackjack/domain/participant/Player.kt
@@ -1,8 +1,7 @@
 package blackjack.domain.participant
 
-import blackjack.domain.BetAmount
-import blackjack.domain.Hand
 import blackjack.domain.card.Card
+import blackjack.domain.hand.Hand
 
 class Player(
     val name: String,
