+++ src/test/kotlin/blackjack/domain/result/MatchResultTest.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.result
 
-import blackjack.domain.BetAmount
+import blackjack.domain.participant.BetAmount
 import blackjack.domain.participant.Dealer
 import blackjack.test.TestObjectGenerator
 import io.kotest.core.spec.style.BehaviorSpec
